requirements.txt
click==6.7
colorama==0.3.9
Django==1.11.7
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
djangorestframework==3.7.3
python-dateutil==2.6.1
python-vuejs==1.0.9
pytz==2017.3
six==1.11.0
sqlparse==0.2.4
